Remove auth+http clients in favor of gordon-gcp
