+++ docs/exceptions.rst
@@ -7,8 +7,4 @@ Any exception that this module could throw.
 
 .. autoclass:: gordon_janitor_gcp.GCPGordonJanitorError
 
-.. autoclass:: gordon_janitor_gcp.GCPAuthError
-
-.. autoclass:: gordon_janitor_gcp.GCPHTTPError
-
 .. autoclass:: gordon_janitor_gcp.GCPConfigError
