+++ src/gordon_janitor_gcp/plugins/reconciler.py
@@ -59,10 +59,10 @@ import logging
 
 import attr
 import zope.interface
+from gordon_gcp.clients import auth
 from gordon_janitor import interfaces
 
 from gordon_janitor_gcp import exceptions
-from gordon_janitor_gcp.clients import auth
 from gordon_janitor_gcp.clients import gdns
